+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.3]
+## Fixed
+- Fixed the issue where component would have kept itself in the array of `parentsHavingFocusedChild` array after removal
+- Further improvements to `autoRestoreFocus` logic to trigger not only on Lead components, but also on Parents that had focused child when being removed. Edge case, normally children are removed first.
+
 # [1.3.2]
 ## Fixed
 - Fixed a bug where parents were not updating their `hasFocusedChild` when new child is created and focused right away